+++ src/types/index.ts
@@ -10,3 +10,4 @@ export * from './UpdateRemoteFunction';
 export * from './ThemeObjectsList';
 export * from './DeepTokensMap';
 export * from './MapValuesToTokensResult';
+export * from './ShowFormOptions';
